backend_upload

+++ plant_village/src/main/java/plant_village/service/PredictionService.java
@@ -1,9 +1,8 @@
 package plant_village.service;
 
-import plant_village.model.Prediction;
-import plant_village.model.User;
-import java.util.List;
-import java.util.Optional;
+import plant_village.model.*;
+import lombok.extern.slf4j.Slf4j;
+import java.util.*;
 
 public interface PredictionService {
     
@@ -26,5 +25,5 @@ public interface PredictionService {
      * Real-time prediction API
      * Processes plant disease prediction with image and sends result via WebSocket
      */
-    Prediction predictPlantDisease(Long userId, Long plantId, String imageBase64, String description);
+    Prediction predictPlantDisease(Integer userId, Integer plantId, String imageBase64, String description);
 }
